Use the current authenticated user for building queries

+++ src/main/java/com/zack/shops/Exception/UserNotFoundException.java
@@ -2,7 +2,7 @@ package com.zack.shops.Exception;
 
 public class UserNotFoundException extends RuntimeException {
 
-	public UserNotFoundException(int id) {
+	public UserNotFoundException(Long id) {
 		super("Could not find User with id : " + id);
 	}
 }
